Dont track doc/tags
doc/tags is a file generated by the :helptags command.
If you track doc/tags and change the documentation, you'd have to
regenerate this file every time and commit it or users would get
warnings about doc/tags having modified content.
The better option is to add the file to the .gitignore (as done by
previous commit) and if it's already being tracked by git, stop tracking
it via `git rm`.